+++ coralogix/resource_coralogix_api_key_test.go
@@ -1,0 +1,71 @@
+package coralogix
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+var (
+	apiKeyResourceName = "coralogix_api_key.test"
+	targetTeam         = "4013254"
+)
+
+func TestApiKeyResource(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testApiKeyResource(),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(apiKeyResourceName, "name", "Test Key 3"),
+					resource.TestCheckResourceAttr(apiKeyResourceName, "owner.team_id", targetTeam),
+					resource.TestCheckResourceAttr(apiKeyResourceName, "active", "true"),
+					resource.TestCheckResourceAttr(apiKeyResourceName, "hashed", "false"),
+				),
+			},
+			{
+				ResourceName:      apiKeyResourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+			{
+				Config: updateApiKeyResource(),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(apiKeyResourceName, "name", "Test Key 5"),
+					resource.TestCheckResourceAttr(apiKeyResourceName, "owner.team_id", targetTeam),
+					resource.TestCheckResourceAttr(apiKeyResourceName, "active", "false"),
+					resource.TestCheckResourceAttr(apiKeyResourceName, "hashed", "false"),
+				),
+			},
+		},
+	})
+}
+
+func testApiKeyResource() string {
+	return strings.Replace(`resource "coralogix_api_key" "test" {
+  name  = "Test Key 3"
+  owner = {
+    team_id : "<TEAM_ID>"
+  }
+  active = true
+  hashed = false
+  roles = ["SCIM"]
+}
+`, "<TEAM_ID>", targetTeam, 1)
+}
+
+func updateApiKeyResource() string {
+	return strings.Replace(`resource "coralogix_api_key" "test" {
+  name  = "Test Key 5"
+  owner = {
+    team_id : "<TEAM_ID>"
+  }
+  active = false
+  hashed = false
+  roles = ["SCIM"]
+}
+`, "<TEAM_ID>", targetTeam, 1)
+}